port cpuj to 32-bit

+++ dbg.h
@@ -7,7 +7,7 @@
 #define DBG_MAX_BREAKPOINTS 16
 
 typedef struct {
-    uint16_t breakpoints[DBG_MAX_BREAKPOINTS]; /* byte addresses */
+    uint32_t breakpoints[DBG_MAX_BREAKPOINTS]; /* byte addresses */
     int      nbreak;
     int      cycles;   /* cycle count since debugger started */
 } dbg_t;
@@ -17,11 +17,11 @@ typedef struct {
  * false if the user quit. */
 bool dbg_run(cpuj_t *cpu, const asm_result_t *prog, dbg_t *dbg);
 
-void dbg_add_breakpoint(dbg_t *dbg, uint16_t addr);
-bool dbg_remove_breakpoint(dbg_t *dbg, uint16_t addr);
+void dbg_add_breakpoint(dbg_t *dbg, uint32_t addr);
+bool dbg_remove_breakpoint(dbg_t *dbg, uint32_t addr);
 void dbg_print_state(const cpuj_t *cpu);
 void dbg_print_regs(const cpuj_t *cpu);
-void dbg_print_mem(const cpuj_t *cpu, uint16_t from, uint16_t to);
-void dbg_print_insn(const cpuj_t *cpu, uint16_t addr);
+void dbg_print_mem(const cpuj_t *cpu, uint32_t from, uint32_t to);
+void dbg_print_insn(const cpuj_t *cpu, uint32_t addr);
 
 #endif /* CPUJ_DBG_H */
